feat: implement dashboard stats and recent tasks components; add stat cards and table view for task management
