+++ src/components/overview/dashboard-stats/stat-card.tsx
@@ -1,0 +1,30 @@
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import type { LucideIcon } from 'lucide-react'
+
+export interface StatCardProps {
+  title: string
+  value: number
+  description: string
+  Icon: LucideIcon
+  color: string
+}
+
+export function StatCard({ title, value, description, Icon, color }: StatCardProps) {
+  return (
+    <Card>
+      <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
+        <CardTitle className="text-xs md:text-sm font-medium text-muted-foreground truncate pr-1">{title}</CardTitle>
+        <div
+          className="h-8 w-8 rounded-full flex items-center justify-center"
+          style={{ backgroundColor: `color-mix(in srgb, ${color} 15%, transparent)` }}
+        >
+          <Icon className="h-4 w-4" style={{ color }} />
+        </div>
+      </CardHeader>
+      <CardContent>
+        <div className="text-xl md:text-2xl font-bold">{value}</div>
+        <p className="text-xs text-muted-foreground mt-1">{description}</p>
+      </CardContent>
+    </Card>
+  )
+}
